requirements.txt
typer>=0.12.0
rich>=13.0.0
httpx>=0.28.0
dnspython>=2.6.0
boto3>=1.34.0
